refactor(genesis)!: consolidate GnDocsExample subcomponents into one folder
Move all eight GnDocsExample* .vue files from individual sibling
folders into a single packages/genesis/src/components/GnDocsExample/
directory. The eight subfolder index.ts files collapse into one
index.ts that re-exports the whole suite.

BREAKING CHANGE: deep imports from the old subfolder paths
(`@paper/genesis/src/components/GnDocsExampleActions` etc.) no longer
resolve. Consumers of the package barrel (`import { X } from
'@paper/genesis'`) are unaffected.
