+++ packages/genesis/src/components/GnDocsExample/GnDocsExampleActions.vue
