+++ packages/genesis/src/components/GnDocsExample/GnDocsExamplePeek.vue
