+++ packages/genesis/src/components/GnDocsExample/index.ts
@@ -1,2 +1,16 @@
 export type { GnDocsExampleProps } from './GnDocsExample.vue'
 export { default as GnDocsExample } from './GnDocsExample.vue'
+export type { GnDocsExampleActionsProps } from './GnDocsExampleActions.vue'
+export { default as GnDocsExampleActions } from './GnDocsExampleActions.vue'
+export type { GnDocsExampleCodeProps } from './GnDocsExampleCode.vue'
+export { default as GnDocsExampleCode } from './GnDocsExampleCode.vue'
+export type { GnDocsExampleDescriptionProps } from './GnDocsExampleDescription.vue'
+export { default as GnDocsExampleDescription } from './GnDocsExampleDescription.vue'
+export type { GnDocsExamplePanelProps } from './GnDocsExamplePanel.vue'
+export { default as GnDocsExamplePanel } from './GnDocsExamplePanel.vue'
+export type { GnDocsExamplePeekProps } from './GnDocsExamplePeek.vue'
+export { default as GnDocsExamplePeek } from './GnDocsExamplePeek.vue'
+export type { GnDocsExamplePreviewProps } from './GnDocsExamplePreview.vue'
+export { default as GnDocsExamplePreview } from './GnDocsExamplePreview.vue'
+export type { GnDocsExampleFile, GnDocsExampleTabsProps } from './GnDocsExampleTabs.vue'
+export { default as GnDocsExampleTabs } from './GnDocsExampleTabs.vue'
