--- packages/genesis/src/components/GnDocsExampleActions/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExampleActionsProps } from './GnDocsExampleActions.vue'
-export { default as GnDocsExampleActions } from './GnDocsExampleActions.vue'
