--- packages/genesis/src/components/GnDocsExampleCode/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExampleCodeProps } from './GnDocsExampleCode.vue'
-export { default as GnDocsExampleCode } from './GnDocsExampleCode.vue'
