--- packages/genesis/src/components/GnDocsExampleDescription/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExampleDescriptionProps } from './GnDocsExampleDescription.vue'
-export { default as GnDocsExampleDescription } from './GnDocsExampleDescription.vue'
