--- packages/genesis/src/components/GnDocsExamplePanel/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExamplePanelProps } from './GnDocsExamplePanel.vue'
-export { default as GnDocsExamplePanel } from './GnDocsExamplePanel.vue'
