--- packages/genesis/src/components/GnDocsExamplePeek/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExamplePeekProps } from './GnDocsExamplePeek.vue'
-export { default as GnDocsExamplePeek } from './GnDocsExamplePeek.vue'
