--- packages/genesis/src/components/GnDocsExamplePreview/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExamplePreviewProps } from './GnDocsExamplePreview.vue'
-export { default as GnDocsExamplePreview } from './GnDocsExamplePreview.vue'
