--- packages/genesis/src/components/GnDocsExampleTabs/index.ts
@@ -1,2 +1,0 @@
-export type { GnDocsExampleFile, GnDocsExampleTabsProps } from './GnDocsExampleTabs.vue'
-export { default as GnDocsExampleTabs } from './GnDocsExampleTabs.vue'
