+++ packages/genesis/src/components/index.ts
@@ -1,8 +1,1 @@
 export * from './GnDocsExample'
-export * from './GnDocsExampleActions'
-export * from './GnDocsExampleCode'
-export * from './GnDocsExampleDescription'
-export * from './GnDocsExamplePanel'
-export * from './GnDocsExamplePeek'
-export * from './GnDocsExamplePreview'
-export * from './GnDocsExampleTabs'
